delete excpetion

+++ course/src/main/java/com/zebodexxsz/course/services/UserService.java
@@ -4,48 +4,52 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.zebodexxsz.course.entities.User;
 import com.zebodexxsz.course.repositories.UserRepository;
+import com.zebodexxsz.course.resources.exceptions.DatabaseException;
+import com.zebodexxsz.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll()
-	{
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
-	public User findById(Long id)
-	{
+
+	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
-		
-		return obj.get();
+
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
-	public User insert(User obj)
-	{
+
+	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
-	public void delete(Long id)
-	{
-		repository.deleteById(id);
+
+	public void delete(Long id) {
+		try {
+			if (repository.findById(id) != null) {
+				repository.deleteById(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
-	public User update(Long id, User obj)
-	{
+
+	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
 		return repository.save(entity);
 	}
-	
-	private void updateData(User entity, User obj)
-	{
+
+	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
